requirements.txt
Flask
firebase-admin
python-dotenv
markdown
Werkzeug